requirements.txt
flask
flask-cors
python-dotenv
google-genai
PyPDF2
numpy
